requirements/base.txt
boto3==1.4.2
botocore==1.4.87
dealer==2.0.5
dj-database-url==0.4.1
Django==1.10.4
django-classy-tags==0.8.0
django-dotenv==1.4.1
django-localflavor==1.3
django-sekizai==0.10.0
django_storages==1.5.1
foundation
googlemaps==2.4.6
gunicorn==19.6.0
jmespath==0.9.0
psycopg2==2.6.2
python-dateutil==2.6.0
python3-memcached==1.51
pytz==2016.10
requests==2.11.1
six==1.10.0
s3transfer==0.1.10
